3DUSE plugins support updated - Visibilite is now a plugin (ok under Windows, Linux and Mac OS X) Under Linux you must recompile and install external/laslib before // from /externals/laslib folder // mkdir build; cd build; cmake .. -DCMAKE_BUILD_TYPE=Release; cmake .. -DCMAKE_BUILD_TYPE=Release; make; sudo make install

+++ src/gui/osg/osgGDAL.cpp
@@ -178,7 +178,7 @@ osg::ref_ptr<osg::Geode> buildOsgGDAL(OGRDataSource* poDS)
 						indices->push_back(cpt++);
 					}
 
-					for(unsigned int k = 0; k < poPG->getNumInteriorRings(); k++)
+					for(unsigned int k = 0; k < (unsigned int)poPG->getNumInteriorRings(); k++)
 					{
 						OGRLinearRing* poLR = poPG->getInteriorRing(k);
